+++ module/api/dice/meta-execute.mjs
@@ -29,8 +29,8 @@ export async function metaExecute({event, actorUUID, action, itemName, multiActi
 		multiAction = parseInt(clickedButton.dataset.multiAction) ?? 0;
 	}
 	//? Check if we are running in Alpha/Beta Testing mode (available via Homebrew/Core Module respectfuly)
-	const alphaTesting = game.settings.get("metanthropes", "metaAlphaTesting");
-	const betaTesting = game.settings.get("metanthropes", "metaBetaTesting");
+	const alphaTesting = metanthropes.utils.metaCheckSetting("homebrew", "metaAlphaTesting");
+	const betaTesting = metanthropes.utils.metaCheckSetting("core", "metaBetaTesting");
 	const actor = await fromUuid(actorUUID);
 	//? Checking if actor has Metapowers that affect the explosive dice
 	const explosiveDice = "x10"; //todo: placeholder for custom explosive dice